Override client-side CacheListenerTest to use correct entry filter
The entry filter used in CacheListenerTest.testFilteredEvents
has been overridden on member-side CacheListenerTest to fix a
NullPointerException in the entry event filter in released
cache-tests:1.1.0 artifact. The same fix is now applied also
on the client-side CacheListenerTest.

(cherry picked from commit 1aa90a0)
